Added canvasHost + accessToken passthru for endpoints
